requirements.txt
streamlit==1.28.0
langchain==0.0.346
langchain-community==0.0.7
groq==0.3.0
chromadb==0.4.15
unstructured==0.10.0
pypdf2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.9.0.post1
tiktoken==0.5.1
numpy==1.26.4
pandas==2.0.0